q1_medical_assistant/requirements.txt
# Core RAG and AI dependencies - Updated for Python 3.11 compatibility
openai>=1.12.0,<2.0.0
langchain>=0.1.0,<0.2.0
langchain-openai>=0.0.5,<0.1.0
langchain-community>=0.0.20,<0.1.0
langchain-core>=0.1.0,<0.2.0
langchain-text-splitters>=0.0.1,<0.1.0
chromadb>=0.4.22,<0.5.0
sentence-transformers>=2.2.2,<3.0.0

# RAGAS evaluation framework - Updated versions
ragas>=0.1.0,<0.2.0
datasets>=2.15.0,<3.0.0
evaluate>=0.4.1,<0.5.0

# Web framework and API - Updated for compatibility
fastapi>=0.104.1,<0.105.0
uvicorn[standard]>=0.24.0,<0.25.0
python-multipart>=0.0.6,<0.1.0
pydantic>=2.5.0,<3.0.0

# Document processing - Updated versions
pypdf2>=3.0.1,<4.0.0
python-docx>=1.1.0,<2.0.0
pymupdf>=1.23.8,<2.0.0
pdfplumber>=0.10.3,<0.11.0

# Data processing and utilities - Updated for Python 3.11
pandas>=2.1.4,<3.0.0
numpy>=1.24.4,<2.0.0
scikit-learn>=1.3.2,<2.0.0
nltk>=3.8.1,<4.0.0

# Monitoring and logging - Updated versions
prometheus-client>=0.19.0,<0.20.0
structlog>=23.2.0,<24.0.0
python-json-logger>=2.0.7,<3.0.0

# HTTP and networking - Updated versions
httpx>=0.25.2,<0.26.0
aiofiles>=23.2.1,<24.0.0
requests>=2.31.0,<3.0.0

# Environment and configuration - Updated versions
python-dotenv>=1.0.0,<2.0.0
pydantic-settings>=2.1.0,<3.0.0

# Testing - Updated versions
pytest>=7.4.3,<8.0.0
pytest-asyncio>=0.21.1,<0.22.0
pytest-cov>=4.1.0,<5.0.0

# Development tools - Updated versions
black>=23.11.0,<24.0.0
isort>=5.12.0,<6.0.0
flake8>=6.1.0,<7.0.0
mypy>=1.7.1,<2.0.0

# Medical and scientific - Updated versions
scipy>=1.11.4,<2.0.0
matplotlib>=3.8.2,<4.0.0
seaborn>=0.13.0,<0.14.0

# Vector operations - Updated version
faiss-cpu>=1.7.4,<2.0.0

# Async support - Updated version
asyncio-mqtt>=0.16.1,<0.17.0

# Security - Updated version
cryptography>=41.0.0,<42.0.0

# Docker support - Updated version
docker>=6.1.3,<7.0.0

# Additional dependencies for compatibility
typing-extensions>=4.8.0,<5.0.0
packaging>=23.0,<24.0
setuptools>=68.0.0,<69.0.0
wheel>=0.41.0,<0.42.0 